ref: support collapseStrategy on Flamegraph
